backend/requirements.txt
dnspython
pyyaml
requests
Django==4.2.11
django-ninja
bs4
celery
redis
psycopg2-binary